refactor(voice): decompose rex/voice_loop.py into rex/voice concern modules (US-REM-028)
rex/voice_loop.py is now a 127-line facade re-exporting from rex/voice/ with rex.voice_loop.<name> kept as the import and monkeypatch surface. Also restores CLAUDE.md and PRD content truncated by a filesystem read race in the previous commit; six source-inspection tests now aggregate the decomposed sources in original order.
