+++ rex/voice/__init__.py
@@ -1,0 +1,4 @@
+"""Voice pipeline modules, decomposed by concern (US-REM-028).
+
+The public surface remains ``rex.voice_loop``; import from there.
+"""
