+++ tests/test_us170_blocking_audit.py
@@ -8,7 +8,22 @@ VOICE_LOOP_SRC = REPO_ROOT / "rex" / "voice_loop.py"
 
 
 def _src() -> str:
-    return VOICE_LOOP_SRC.read_text(encoding="utf-8")
+    parts = [VOICE_LOOP_SRC.read_text(encoding="utf-8")]
+    # Original rex/voice_loop.py concern order (US-REM-028 decomposition).
+    for _name in (
+        "optional_imports.py",
+        "audio_utils.py",
+        "_types.py",
+        "transcripts.py",
+        "microphone.py",
+        "acknowledgement.py",
+        "stt.py",
+        "tts.py",
+        "loop.py",
+        "builder.py",
+    ):
+        parts.append((REPO_ROOT / "rex" / "voice" / _name).read_text(encoding="utf-8"))
+    return "\n".join(parts)
 
 
 # ── No time.sleep in hot path ─────────────────────────────────────────────────
