+++ tests/test_us171_wake_ack.py
@@ -11,7 +11,22 @@ PERF_DOC = REPO_ROOT / "docs" / "performance-baseline.md"
 
 
 def _src() -> str:
-    return VOICE_LOOP_SRC.read_text(encoding="utf-8")
+    parts = [VOICE_LOOP_SRC.read_text(encoding="utf-8")]
+    # Original rex/voice_loop.py concern order (US-REM-028 decomposition).
+    for _name in (
+        "optional_imports.py",
+        "audio_utils.py",
+        "_types.py",
+        "transcripts.py",
+        "microphone.py",
+        "acknowledgement.py",
+        "stt.py",
+        "tts.py",
+        "loop.py",
+        "builder.py",
+    ):
+        parts.append((REPO_ROOT / "rex" / "voice" / _name).read_text(encoding="utf-8"))
+    return "\n".join(parts)
 
 
 # ── Source-level checks ────────────────────────────────────────────────────────
